date formatting change for Safari
Someone just alerted me that the Maptime Chicago "upcoming events" box was showing "TBD" despite an upcoming event this week. I looked into it and, while Chrome and Firefox were working fine, Safari was indeed showing "TBD." 

After some poking around, it seems that Safari does not like dashes is a-ok with slashes. The proposed change is to go from `yyyy-MM-dd` to `yyyy/MM/dd`. I've made this change to the Maptime Chicago repository and it is now working in Chrome, FF and Safari.